# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel perfume api",
        "description": "L5 Swagger OpenApi ",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000"
        }
    ],
    "paths": {
...
        "/api/admin/category": {
            "get": {
                "summary": "گرفتن تمام کتگوری های موجود برای ادمین",
                "operationId": "e73c2ef1d12b01d27ea546e313b60f71",
                "parameters": [
                    {
                        "name": "access token",
                        "in": "path",
                        "description": "احراز هویت با توکن",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/AdminCategory"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "unAuthorized"
                    }
                }
            },
            "post": {
                "summary": "اضافه کردن کتگوری ",
                "operationId": "cd540e7d359e4a8093b44cde91faf28a",
                "parameters": [
                    {
                        "name": "access token",
                        "in": "path",
                        "description": "احراز هویت با توکن",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "perfume",
                                            "watch"
                                        ]
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "slug": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "name": "woody orbal category",
                                    "type": "perfume",
                                    "description": "مخصوص فصل های گرم ",
                                    "slug": "woody-orbal"
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "unAuthorized"
                    }
                }
            }
        },
        "/api/admin/category/{slug}": {
            "get": {
                "summary": "گرفتن اطلاعات مربوط به یک کتگوری خاص برای ادمین",
                "operationId": "76f6fe336d7b7d4d7ffe4e08743a4cc2",
                "parameters": [
                    {
                        "name": "access token",
                        "in": "path",
                        "description": "احراز هویت با توکن",
                        "required": true
                    },
                    {
                        "name": "اسلاگ در یو ار ال",
                        "in": "path",
                        "description": "اسلاگ کتگوری مربوط در یو ار ال",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/AdminFullCategory"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "unAuthorized"
                    }
                }
            }
        },
        "/api/admin/category/{category:slug}": {
            "put": {
                "summary": "تغییر کتگوری موجود",
                "operationId": "92a159f7f9839eaae6d80c2115b4eb6c",
                "parameters": [
                    {
                        "name": "access token",
                        "in": "path",
                        "description": "احراز هویت با توکن",
                        "required": true
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "اسلاگ سوال در روت",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "perfume",
                                            "watch"
                                        ]
                                    },
                                    "description": {
                                        "type": "string"
                                    },
                                    "slug": {
                                        "type": "string"
                                    }
                                },
                                "type": "object",
                                "example": {
                                    "question": "سوال مورد نظر",
                                    "answer": "جواب مورد نظر "
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "unAuthorized"
                    }
                }
            },
            "delete": {
                "summary": "حذف کتگوری موجود",
                "operationId": "b9b93d6503cb70a7c4568dd43f1349af",
                "parameters": [
                    {
                        "name": "access token",
                        "in": "path",
                        "description": "احراز هویت با توکن",
                        "required": true
                    },
                    {
                        "name": "id",
                        "in": "path",
                        "description": "اسلاگ سوال در روت",
                        "required": true
                    }
                ],
                "responses": {
                    "202": {
                        "description": "OK"
                    },
                    "403": {
                        "description": "unAuthorized"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "AdminCategory": {
                "title": "  کتگوری ادمین",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "type": {
                        "type": "string"
                    },
                    "slug": {
                        "type": "string"
                    },
                    "is_active": {
                        "type": "string",
                        "default": "فعال"
                    }
                },
                "type": "object"
            },
            "AdminFullCategory": {
                "title": "   یک کتگوری برای ادمین",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "id": {
                        "type": "integer"
                    },
                    "type": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "slug": {
                        "type": "string"
                    },
                    "is_active": {
                        "type": "string",
                        "default": "فعال"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}